requirements.txt
mysqlclient >=1.4.2
PyYAML >=3.12
